+++ qiling/hw/gpio/stm32f1xx_gpio.py
@@ -40,3 +40,59 @@ class STM32F1xxGpio(QlPeripheral, GpioHooks):
         GpioHooks.__init__(self, ql, 16)
 
         self.gpio = self.struct()
+
+    @QlPeripheral.monitor()
+    def read(self, offset: int, size: int) -> int:
+        if offset == self.struct.BSRR.offset:
+            return 0x00
+        
+        buf = ctypes.create_string_buffer(size)
+        ctypes.memmove(buf, ctypes.addressof(self.gpio) + offset, size)
+        return int.from_bytes(buf.raw, byteorder='little')
+
+    @QlPeripheral.monitor()
+    def write(self, offset: int, size: int, value: int):
+        if offset == self.struct.IDR.offset:
+            return
+
+        if offset == self.struct.BSRR.offset:
+            for i in range(32):
+                if ((value >> i) & 1) == 0:
+                    continue
+                if i < 16:   
+                    self.set_pin(i)
+                else:
+                    self.reset_pin(i - 16)                    
+            
+            return
+        
+        if offset == self.struct.ODR.offset:            
+            for i in range(16):
+                new_bit = (value >> i) & 1
+                old_bit = (self.gpio.ODR >> i) & 1                
+
+                if new_bit !=  old_bit:
+                    if new_bit:
+                        self.set_pin(i)                        
+                    else:
+                        self.reset_pin(i)                        
+            
+            return    
+        
+        data = (value).to_bytes(size, 'little')
+        ctypes.memmove(ctypes.addressof(self.gpio) + offset, data, size) 
+
+    def set_pin(self, i):
+        self.ql.log.debug(f'[{self.label}] Set P{self.label[-1].upper()}{i}')
+        
+        self.gpio.ODR |= 1 << i        
+        self.call_hook_set(i)
+    
+    def reset_pin(self, i):
+        self.ql.log.debug(f'[{self.label}] Reset P{self.label[-1].upper()}{i}')
+        
+        self.gpio.ODR &= ~(1 << i)
+        self.call_hook_reset(i)
+        
+    def pin(self, index):
+        return (self.gpio.ODR >> index) & 1
